+++ toolkit/cli/smart-contracts-commands/src/reserve.rs
@@ -13,6 +13,7 @@ use partner_chains_cardano_offchain::{
 	},
 };
 use sidechain_domain::{AssetId, ScriptHash, UtxoId};
+use std::num::NonZero;
 
 #[derive(Clone, Debug, clap::Subcommand)]
 #[allow(clippy::large_enum_variant)]
@@ -208,12 +209,9 @@ pub struct ReleaseReserveCmd {
 	/// Reference UTXO containing the V-Function script
 	#[arg(long, short('r'))]
 	reference_utxo: UtxoId,
-	/// Encoded asset id in form <policy_id_hex>.<asset_name_hex>.
+	/// Amount of reserve tokens to be released to the illiquid supply.
 	#[arg(long)]
-	token: AssetId,
-	/// Current value of the V-Function
-	#[arg(long)]
-	amount: u64,
+	amount: NonZero<u64>,
 }
 
 impl ReleaseReserveCmd {
@@ -221,7 +219,7 @@ impl ReleaseReserveCmd {
 		let payment_key = self.payment_key_file.read_key()?;
 		let ogmios_client = client_for_url(&self.common_arguments.ogmios_url).await?;
 		let _ = release_reserve_funds(
-			TokenAmount { token: self.token, amount: self.amount },
+			self.amount,
 			self.genesis_utxo,
 			self.reference_utxo,
 			&payment_key,